Add newscorp files

+++ newscorp/local.zsh
@@ -13,9 +13,51 @@ webapp() {
 
 phpunit_plugin () {
   if [[ $# != 1 ]]; then echo "usage: phpunit_plugin plugin"; return; fi
+  if [[ $(is_plugin $1) != 1 ]]; then echo "$1 is not a valid plugin!"; return; fi
   set -x
   pth=/srv/www/wp-content/themes/vip/newscorpau-plugins/$1/phpunit.xml
   docker-compose run --rm ci phpunit -c "$pth" "${@:2}"
+}
+
+phpcs_plugin() {
+  if [[ $# != 1 ]]; then echo "usage: phpcs_plugin plugin"; return; fi
+  if [[ $(is_plugin $1) != 1 ]]; then echo "$1 is not a valid plugin!"; return; fi
+  pth=/srv/www/wp-content/themes/vip/newscorpau-plugins/$1
+  docker-compose run --rm ci phpcs -p -s -v --standard="$pth/phpcs.ruleset.xml" --ignore={tests,vendor}/\* $pth
+  if [[ $? != 0 ]]; then
+    echo "ERROR. Check above"
+  else
+    echo "OKAY."
+  fi
+}
+
+phpcs_plugin_branch() {
+  if [[ $# != 2 ]]; then echo "usage: phpcs_plugin plugin branch"; return; fi
+  if [[ $(is_plugin $1) != 1 ]]; then echo "$1 is not a valid plugin!"; return; fi
+  co_test.py master
+  plugin=$1
+  branch=$2
+  pth=/srv/www/wp-content/themes/vip/newscorpau-plugins/$plugin
+  localpth=$VIP/www/wp-content/themes/vip/newscorpau-plugins/$plugin
+  files=($(git -C "$localpth" diff master $branch --name-only))
+  git -C "$localpth" checkout $branch
+  for file in $files; do
+    echo $file
+    docker-compose run --rm ci phpcs -p -s -v --standard="$pth/phpcs.ruleset.xml" --ignore={tests,vendor}/\* $pth/$file
+    if [[ $? != 0 ]]; then
+      echo "FAILED."
+      return
+    fi
+  done
+  echo "OKAY."
+}
+
+is_plugin() {
+  if [[ -d "$VIP/www/wp-content/themes/vip/newscorpau-plugins/$1" ]]; then
+    echo 1
+  else
+    echo 0
+  fi
 }
 
 gcot() {
